// web/src/registry.generated.ts
// Generated by scripts/generate_indicator_keys.py; do not edit by hand.
export const INDICATOR_KEYS = [
  "btc_daily_close",
  "btc_rsi",
  "btc_funding_rate",
  "btc_mvrv",
  "btc_active_addresses",
  "btc_exchange_flow",
  "btc_open_interest",
  "btc_long_short_ratio",
  "btc_taker_ratio",
  "btc_ema_20",
  "btc_ema_50",
  "btc_ema_200",
  "btc_atr",
  "btc_bollinger_upper",
  "btc_bollinger_middle",
  "btc_bollinger_lower",
  "btc_obv",
  "btc_macd",
  "btc_stochrsi",
  "eth_rsi",
  "eth_funding_rate",
  "eth_mvrv",
  "eth_active_addresses",
  "eth_exchange_flow",
  "eth_open_interest",
  "eth_long_short_ratio",
  "eth_taker_ratio",
  "eth_ema_20",
  "eth_ema_50",
  "eth_ema_200",
  "eth_atr",
  "eth_bollinger_upper",
  "eth_bollinger_middle",
  "eth_bollinger_lower",
  "eth_obv",
  "eth_macd",
  "eth_stochrsi",
  "sol_rsi",
  "sol_funding_rate",
  "sol_active_addresses",
  "sol_staking",
  "sol_open_interest",
  "sol_long_short_ratio",
  "sol_taker_ratio",
  "sol_ema_20",
  "sol_ema_50",
  "sol_ema_200",
  "sol_atr",
  "sol_bollinger_upper",
  "sol_bollinger_middle",
  "sol_bollinger_lower",
  "sol_obv",
  "sol_macd",
  "sol_stochrsi",
  "bnb_rsi",
  "bnb_funding_rate",
  "bnb_mvrv",
  "bnb_active_addresses",
  "bnb_open_interest",
  "bnb_long_short_ratio",
  "bnb_taker_ratio",
  "bnb_ema_20",
  "bnb_ema_50",
  "bnb_ema_200",
  "bnb_atr",
  "bnb_bollinger_upper",
  "bnb_bollinger_middle",
  "bnb_bollinger_lower",
  "bnb_obv",
  "bnb_macd",
  "bnb_stochrsi",
  "macro_vixcls",
  "macro_dff",
  "macro_t10y2y",
  "macro_dfii10",
  "macro_dtwexbgs",
  "macro_cpiaucsl",
  "macro_m2sl",
  "spot_etf_net_flow",
  "stablecoin_supply",
  "fear_greed_index"
] as const;

export type IndicatorKey = (typeof INDICATOR_KEYS)[number];